requirements.txt
crew
python-dotenv